Move tail filter out of timeseries.py
