+++ libs/datasets/tail_filter.py
@@ -1,0 +1,124 @@
+from typing import Any, ClassVar, List, Mapping, Tuple
+import math
+import dataclasses
+
+import pandas as pd
+from covidactnow.datapublic.common_fields import FieldName
+from covidactnow.datapublic.common_fields import PdFields
+
+from libs.datasets import timeseries
+
+AnnotationType = timeseries.AnnotationType
+AnnotationField = timeseries.AnnotationField
+
+
+@dataclasses.dataclass
+class TailFilter:
+    annotations: List[Mapping[str, Any]] = dataclasses.field(default_factory=list)
+
+    # Counts that track what the filter has done.
+    skipped_too_short: int = 0
+    skipped_na_mean: int = 0
+    all_good: int = 0
+    truncated: int = 0
+    long_truncated: int = 0
+
+    # Trust values 2 - 4 weeks ago to have reasonable diffs. Access series_in.iat[-15] to
+    # series_in.iat[-28] (inclusive).
+    TRUSTED_DATES_OLDEST: ClassVar[int] = -28
+    TRUSTED_DATES_NEWEST: ClassVar[int] = -15
+
+    # Possibly drop series_in.iat[-1] to series_in.iat[-14] (inclusive).
+    FILTER_DATES_OLDEST: ClassVar[int] = -14
+    # If this many real values are dropped use the more severe AnnotationType.
+    COUNT_OBSERVATION_LONG: ClassVar[int] = 8
+
+    @staticmethod
+    def run(
+        dataset: timeseries.MultiRegionDataset, fields: List[FieldName]
+    ) -> Tuple["TailFilter", timeseries.MultiRegionDataset]:
+        """Returns a dataset with recent data that looks bad removed from cumulative fields."""
+        timeseries_wide_dates = dataset.timeseries_wide_dates()
+
+        fields_mask = timeseries_wide_dates.index.get_level_values(PdFields.VARIABLE).isin(fields)
+        to_filter = timeseries_wide_dates.loc[pd.IndexSlice[:, fields_mask], :]
+        not_filtered = timeseries_wide_dates.loc[pd.IndexSlice[:, ~fields_mask], :]
+
+        tail_filter = TailFilter()
+        filtered = to_filter.apply(tail_filter._filter_one_series, axis=1)
+
+        merged = pd.concat([not_filtered, filtered])
+        timeseries_wide_variables = merged.stack().unstack(PdFields.VARIABLE).sort_index()
+
+        return tail_filter, dataclasses.replace(dataset, timeseries=timeseries_wide_variables)
+
+    def annotations_as_dataframe(self):
+        return pd.DataFrame(self.annotations)
+
+    def _filter_one_series(self, series_in: pd.Series) -> pd.Series:
+        """Filters one timeseries of cumulative values. This is a method so self can be used to
+        store side outputs.
+
+        Args:
+            series_in: a timeseries of float values, with a sorted DatetimeIndex
+            """
+        if len(series_in) < -TailFilter.TRUSTED_DATES_OLDEST:
+            self.skipped_too_short += 1
+            return series_in
+
+        diff = series_in.diff()
+        mean = diff.iloc[
+            TailFilter.TRUSTED_DATES_OLDEST : TailFilter.TRUSTED_DATES_NEWEST + 1
+        ].mean()
+        if pd.isna(mean):
+            self.skipped_na_mean += 1
+            return series_in
+        # Pick a bound used to determine if the diff of recent values are reasonable. For
+        # now values less than 100th of the mean are considered stalled. The variance could be
+        # used to create a tighter bound but the simple threshold seems to work for now.
+        # TODO(tom): Experiment with other ways to calculate the bound.
+        threshold = math.floor(mean / 100)
+
+        # Go backwards in time, starting with the most recent observation. Stop as soon as there
+        # is an observation with diff that looks reasonable / not stalled / over the threshold.
+        # If no such diff is found after reading iat[FILTER_DATES_OLDEST] the 'else' is taken.
+        count_observation_diff_under_threshold = 0
+        for i in range(-1, TailFilter.FILTER_DATES_OLDEST - 1, -1):  # range stop is exclusive
+            if pd.isna(diff.iat[i]):
+                continue
+            if diff.iat[i] >= threshold:
+                truncate_at = i + 1
+                break
+            else:
+                count_observation_diff_under_threshold += 1
+        else:
+            # Reached end of loop without finding a diff >= threshold.
+            truncate_at = TailFilter.FILTER_DATES_OLDEST
+        if count_observation_diff_under_threshold == 0:
+            self.all_good += 1
+            return series_in
+        else:
+            # series_in.iat[truncate_at] is the first value *not* returned
+            assert TailFilter.FILTER_DATES_OLDEST <= truncate_at <= -1
+            if count_observation_diff_under_threshold < TailFilter.COUNT_OBSERVATION_LONG:
+                annotation_type = AnnotationType.CUMULATIVE_TAIL_TRUNCATED
+                self.truncated += 1
+            else:
+                annotation_type = AnnotationType.CUMULATIVE_LONG_TAIL_TRUNCATED
+                self.long_truncated += 1
+            # Currently one annotation is created per series. Maybe it makes more sense to add
+            # one for each dropped observation / real value?
+            # https://github.com/covid-projections/covid-data-model/pull/855#issuecomment-747698288
+            self.annotations.append(
+                {
+                    AnnotationField.LOCATION_ID: series_in.name[0],
+                    AnnotationField.VARIABLE: series_in.name[1],
+                    AnnotationField.TYPE: annotation_type,
+                    AnnotationField.COMMENT: f"Removed {count_observation_diff_under_threshold} observations that look "
+                    f"suspicious compared to mean diff of {mean:.1f} a few weeks ago.",
+                    AnnotationField.DATE: series_in.index[truncate_at - 1],
+                }
+            )
+            # Using integer position indexing where the upper bound is exclusive, like regular
+            # Python indexing and unlike Pandas label (`loc` and `at`) indexing.
+            return series_in.iloc[:truncate_at]
